% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distspec-package.R
\docType{package}
\name{distspec-package}
\alias{distspec}
\alias{distspec-package}
\title{distspec: Probability Distributions with Certain or Uncertain Parameters}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Represents probability distributions with fixed or potentially uncertain parameters, with tools to discretise, convolve, sample from and summarise them.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://epiforecasts.io/distspec/}
  \item Report bugs at \url{https://github.com/epiforecasts/distspec/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Funk \email{sebastian.funk@lshtm.ac.uk} (\href{https://orcid.org/0000-0002-2842-3406}{ORCID})

Authors:
\itemize{
  \item James M. Azam \email{james.azam@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5782-7330}{ORCID})
  \item Sam Abbott \email{sam.abbott@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-8057-8037}{ORCID})
}

}
\keyword{internal}
